restructure project and add authUser
