+++ src/containers/authUser/authUser.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const AuthUser = () => {
+  return(  
+    <div>
+      АВТОРИЗАЦИЯЯЯЯ
+    </div>
+  )  
+}
+
+export default AuthUser;
