+++ src/main.js
@@ -1,8 +1,11 @@
 import React from "react";
 import { Router, Route } from "react-router";
 import "./main.css";
-import Header from "./components/header/header";
-import Layout from "./components/layout/layout";
+import Header from "./containers/header/header";
+import Layout from "./containers/layout/layout";
+import AuthUser from "./containers/authUser/authUser"
+import { Provider } from 'react-redux'
+import { store } from './store/configureStore'
 
 const createHistory = require("history").createBrowserHistory;
 
@@ -10,17 +13,20 @@ const history = createHistory();
 
 const Main = () => {
   return (
-    <Router history={history}>
-      <div className="grid-container">
-        <div className="header">
-          <Route path="/" component={Header} />
+    <Provider store={store}>
+      <Router history={history}>
+        <div className="grid-container">
+          <div className="header">
+            <Header />
+          </div>
+          <div className="content">
+            <Route path="/auth/login" component={AuthUser} />
+            <Route path="/" component={Layout} />
+          </div>
+          <div className="footer"> exceed-team 2019 </div>
         </div>
-        <div className="content">
-          <Route path="/" component={Layout} />
-        </div>
-        <div className="footer"> exceed-team 2019 </div>
-      </div>
-    </Router>
+      </Router>
+    </Provider>
   );
 };
