+++ src/store/configureStore.js
@@ -1,0 +1,3 @@
+import { createStore } from 'redux'
+
+export const store = createStore(() => {}, {})
